Rename Vue component
